Added dictionary to fix labeling.
The K Means clustering method will correctly classify each student, however, sometimes, (due to the randomness used in the algorithm) the index of the two clusters would switch.  For example, every time it is run, it will always say that students 22 and 23 were in the same class, however, sometimes it will label it as class "1" and sometimes as class "0".  A dictionary has been used to standardize what the classes are called so that they are consistent with the way that the initial dataset is labeled.  This is done by calling the class containing student 0 (the first instructor) "class 0" and the other class which contains student 33 (the second instructor) "class 1".